Restructure into single target to simplify documentation and avoiding rexports
