+++ Plugins/BenchmarkTool/BenchmarkTool+Export.swift
@@ -8,7 +8,7 @@
 // http://www.apache.org/licenses/LICENSE-2.0
 //
 
-import BenchmarkSupport
+import Benchmark
 import DateTime
 import ExtrasJSON
 import SystemPackage
