+++ Sources/Benchmark/BenchmarkMetric+Defaults.swift
