+++ Sources/Benchmark/Documentation.docc/BenchmarkMetric.md
@@ -1,4 +1,4 @@
-# ``BenchmarkSupport/BenchmarkMetric``
+# ``Benchmark/BenchmarkMetric``
 
 ## Topics
