+++ Sources/Benchmark/Documentation.docc/BenchmarkTimeUnits.md
@@ -1,4 +1,4 @@
-# ``BenchmarkSupport/BenchmarkTimeUnits``
+# ``Benchmark/BenchmarkTimeUnits``
 
 ## Topics
