+++ Sources/Benchmark/MallocStats/MallocStats.swift
@@ -12,7 +12,7 @@
 #if swift(>=5.8)
     @_documentation(visibility: internal)
 #endif
-public struct MallocStats {
+internal struct MallocStats {
     var mallocCountTotal: Int = 0 /// total number of mallocs done
     var mallocCountSmall: Int = 0 /// number of small mallocs (as defined by jemalloc)
     var mallocCountLarge: Int = 0 /// number of large mallocs (as defined by jemalloc)
