+++ Sources/Statistics/Statistics.swift
@@ -13,6 +13,9 @@ import Histogram
 import Numerics
 
 /// A type that provides distribution / percentile calculations of latency measurements
+#if swift(>=5.8)
+@_documentation(visibility: internal)
+#endif
 public final class Statistics: Codable {
     public static let defaultMaximumMeasurement = 1_000_000_000 // 1 second in nanoseconds
     public static let defaultPercentilesToCalculate = [0.0, 25.0, 50.0, 75.0, 90.0, 99.0, 100.0]
