+++ Tests/BenchmarkTests/BenchmarkTests.swift
@@ -9,7 +9,7 @@
 //
 
 @testable import Benchmark
-@testable import BenchmarkSupport
+@testable import Benchmark
 import XCTest
 
 final class BenchmarkTests: XCTestCase {
